fix(web): keep agents root on the lazy shell
