+++ docs/todo.md
@@ -5,7 +5,7 @@ Active backlog only. Keep this file small and current.
 - [ ] Verify explicit node startup boundaries after merge: `server.role = "node"` should boot internal gRPC only, skip main-only startup side effects (`ensure_root`, push/VAPID initialization), and fail fast when `server.node_id` is missing or `internal_grpc.enabled` is false. Record focused notes and push/PR CI run IDs in `docs/journal/2026-04-05-node-mode-startup-boundary.md`.
 - [ ] Extend shared-thread Team message sends to honor an explicit server-side `idempotency_key` on `POST /api/teams/:team_id/tasks/:task_id/messages` so repeated submits dedupe across tabs/retries, not just inside one page session. Record follow-up notes in `docs/journal/2026-04-05-team-channel-send-feedback-and-visibility.md`.
 - [ ] **Design System & Tailwind Optimization**: Migrate hardcoded hex colors (e.g., `#f7f7f5`, `#edf2ff`) from `tailwind_classes.ts` into `tailwind.config.cjs` as semantic theme tokens (e.g., `notion-bg-subtle`, `notion-bubble-user`). Reduce redundancy by using composed Tailwind classes for shared bubble patterns.
-- [ ] **Frontend Architecture Refactoring**: Continue decompressing `acp_conversation.tsx` (~2800 LOC). `MarkdownBubble`, `ThinkingBubble`, and `PlanBubble` should live under `src/components/bubbles/`, while the remaining `ToolCallBubble` and tool payload/rendering blocks still need to be split into smaller compound components with stable boundaries.
+- [ ] **Frontend Architecture Refactoring**: Continue shrinking the remaining heavy ACP workbench files after the recent splits: keep decomposing `web/src/app.tsx`, `web/src/components/agents_workbench.tsx`, and the residual tool payload/rendering path so conversation/runtime wiring, payload renderers, and route-shell concerns land in smaller stable modules.
 - [ ] **Framework Integration & UI Standards**: Standardize on Mantine primitives (`UnstyledButton`, `Box`, `ActionIcon`) as base components instead of raw HTML tags to leverage built-in A11y and focus management while styling with Tailwind. Bridge Mantine theme variables (spacing, radius) more deeply into the Tailwind utility layer.
 - [ ] **Performance & UX Hardening**: Evaluate advanced windowing solutions like `react-virtuoso` for the ACP conversation list to ensure smooth scrolling and reliable "stick-to-bottom" behavior in extremely long histories. Implement a `ConversationContext` to manage cross-cutting UI states (e.g., "Collapse All", focus jumping) without deep prop drilling.
 - [ ] Continue consolidating the web workbench onto shared UI primitives: migrate the remaining Team panels, dock/toolbars, and row/list affordances onto the new `web/src/ui/primitives.tsx` layer so repeated Tailwind strings stop drifting independently. Record focused notes and follow-up validation in `docs/journal/2026-04-04-ui-primitives-bootstrap.md`.
